try2  soal_kuis

+++ database/migrations/2024_04_07_144603_create_hasil_kuis_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('users_id');
             $table->integer('materi_id');
-            $table->decimal('skor', 5);
+            $table->integer('skor');
             $table->dateTime('created_at')->useCurrent();
             $table->dateTime('updated_at')->nullable()->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
         });
